Use 32-bit integers instead of 16-bit

+++ pykonal/constants.pyx
@@ -4,6 +4,6 @@ Package-level constants.
 import numpy as np
 
 DTYPE_REAL = np.float64
-DTYPE_UINT = np.uint16
-DTYPE_INT  = np.int16
+DTYPE_UINT = np.uint32
+DTYPE_INT  = np.int32
 DTYPE_BOOL = np.bool
